requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
asyncpg==0.20.1
bitstruct==8.9.0
cbitstruct==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-redis-cache==2.1.1
hexdump==3.3
hiredis==1.0.1
idna==2.9
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2019.3
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.8
